requirements.txt
pillow
opencv-python-headless
tk
